ingestion/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-pubsub==2.18.4
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

